pyproject.toml
[project]
name = "fastapi-all-that-metadata"
version = "0.1.0"
description = "A basic FastAPI web service"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "Pillow>=10.1.0",
    "piexif>=1.1.3",
    "pillow-heif>=0.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
